refactor: relocate layout (API)
